Moved viewmodels to separate files
