+++ Examples/SampleCat/SampleCat/Views/Offerings/OfferingViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  OfferingViewModel.swift
+//  SampleCat
+//
+//  Created by Hidde van der Ploeg on 7/4/25.
+//
+import SwiftUI
+
+struct OfferingViewModel: Identifiable, Equatable {
+    var id: String { identifier }
+    let identifier: String
+    let packages: [PackageViewModel]
+}
